+++ equalsverifier-core/src/main/java/nl/jqno/equalsverifier/internal/checkers/MapEntryHashCodeRequirementChecker.java
@@ -23,18 +23,21 @@ public class MapEntryHashCodeRequirementChecker<T> implements Checker {
     @Override
     public void check() {
         if (Map.Entry.class.isAssignableFrom(config.getType())) {
-            Map.Entry<?, ?> e = valueProvider.<Map.Entry<?, ?>>provideOrThrow(config.getTypeTag()).getRed();
+            Map.Entry<?, ?> e = valueProvider.<Map.Entry<?, ?>>provideOrThrow(config.getTypeTag()).red();
 
             int expectedHashCode = Objects.hashCode(e.getKey()) ^ Objects.hashCode(e.getValue());
             int actualHashCode = config.getCachedHashCodeInitializer().getInitializedHashCode(e);
 
             Formatter f = Formatter
                     .of(
-                        "Map.Entry: hashCode for\n  %%\nshould be %% but was %%.\n"
-                                + "The hash code of a map entry e is defined as:\n"
-                                + "    (e.getKey()==null ? 0 : e.getKey().hashCode()) ^ (e.getValue()==null ? 0 : e.getValue().hashCode())\n"
-                                + "or, using Java 8 API:\n"
-                                + "    java.util.Objects.hashCode(e.getKey()) ^ java.util.Objects.hashCode(e.getValue())",
+                        """
+                                Map.Entry: hashCode for
+                                  %%
+                                should be %% but was %%.
+                                The hash code of a map entry e is defined as:
+                                    (e.getKey()==null ? 0 : e.getKey().hashCode()) ^ (e.getValue()==null ? 0 : e.getValue().hashCode())
+                                or, using Java 8 API:
+                                    java.util.Objects.hashCode(e.getKey()) ^ java.util.Objects.hashCode(e.getValue())""",
                         e,
                         expectedHashCode,
                         actualHashCode);