+++ equalsverifier-core/src/main/java/nl/jqno/equalsverifier/internal/checkers/fieldchecks/NullPointerExceptionFieldCheck.java
@@ -82,12 +82,9 @@ public class NullPointerExceptionFieldCheck<T> implements FieldCheck<T> {
     }
 
     private void abstractMethodErrorThrown(String method, Field field, AbstractMethodError e) {
-        Formatter f = Formatter
-                .of(
-                    "Abstract delegation: %% throws AbstractMethodError when field %% is null.\n"
-                            + "Suppress Warning.NULL_FIELDS to disable this check.",
-                    method,
-                    field.getName());
+        Formatter f = Formatter.of("""
+                Abstract delegation: %% throws AbstractMethodError when field %% is null.
+                Suppress Warning.NULL_FIELDS to disable this check.""", method, field.getName());
         fail(f, e);
     }