+++ equalsverifier-core/src/main/java/nl/jqno/equalsverifier/internal/reflection/ClassProbe.java
@@ -4,9 +4,6 @@ import static nl.jqno.equalsverifier.internal.util.Rethrow.rethrow;
 
 import java.lang.reflect.Modifier;
 
-import nl.jqno.equalsverifier.internal.versionspecific.RecordsHelper;
-import nl.jqno.equalsverifier.internal.versionspecific.SealedTypesHelper;
-
 /**
  * Provides read-only reflective access to a class.
  */
@@ -41,7 +38,7 @@ public final class ClassProbe<T> {
      * @return true if T is a Java Record.
      */
     public boolean isRecord() {
-        return RecordsHelper.isRecord(type);
+        return type.isRecord();
     }
 
     /**
@@ -50,7 +47,7 @@ public final class ClassProbe<T> {
      * @return true if T is a sealed class
      */
     public boolean isSealed() {
-        return SealedTypesHelper.isSealed(type);
+        return type.isSealed();
     }
 
     /**