+++ equalsverifier-core/src/test/java/nl/jqno/equalsverifier/integration/extended_contract/RecordsTest.java
@@ -1,7 +1,6 @@
 package nl.jqno.equalsverifier.integration.extended_contract;
 
 import static nl.jqno.equalsverifier.internal.testhelpers.Util.defaultEquals;
-import static org.assertj.core.api.Assertions.assertThat;
 
 import java.io.Serializable;
 import java.util.Objects;
@@ -9,16 +8,9 @@ import java.util.Objects;
 import nl.jqno.equalsverifier.EqualsVerifier;
 import nl.jqno.equalsverifier.Warning;
 import nl.jqno.equalsverifier.internal.testhelpers.ExpectedException;
-import nl.jqno.equalsverifier.internal.versionspecific.RecordsHelper;
 import org.junit.jupiter.api.Test;
 
 class RecordsTest {
-
-    @Test
-    void sanityCheckForRecordsHelper() {
-        assertThat(RecordsHelper.isRecord(SimpleRecord.class)).isTrue();
-        assertThat(RecordsHelper.isRecord(String.class)).isFalse();
-    }
 
     @Test
     void succeed_whenClassIsARecord() {
@@ -157,34 +149,43 @@ class RecordsTest {
     }
 
     record BrokenInvariantIntFieldRecord(int intField, String stringField) {
+
         public BrokenInvariantIntFieldRecord(int intField, String stringField) {
             this.intField = intField + 1;
             this.stringField = stringField;
         }
+
     }
 
     record BrokenInvariantStringFieldRecord(int intField, String stringField) {
+
         public BrokenInvariantStringFieldRecord(int intField, String stringField) {
             this.intField = intField;
             this.stringField = stringField + "x";
         }
+
     }
 
     record BrokenInvariantBothRecord(int intField, String stringField) {
+
         public BrokenInvariantBothRecord(int intField, String stringField) {
             this.intField = intField + 1;
             this.stringField = stringField + "x";
         }
+
     }
 
     record EqualsRecord(int i, String s) {
+
         @Override
         public boolean equals(Object obj) {
             return defaultEquals(this, obj);
         }
+
     }
 
     record NotAllFieldsRecord(int i, String s) {
+
         @Override
         public boolean equals(Object obj) {
             if (!(obj instanceof NotAllFieldsRecord)) {
@@ -197,6 +198,7 @@ class RecordsTest {
         public int hashCode() {
             return i;
         }
+
     }
 
     record ThrowingConstructorRecord(int i, String s) {
@@ -212,6 +214,7 @@ class RecordsTest {
     }
 
     record ThrowingAccessorRecord(int i, String s) {
+
         public ThrowingAccessorRecord(int i, String s) {
             this.i = i;
             this.s = s + "x";
@@ -220,9 +223,11 @@ class RecordsTest {
         public int i() {
             throw new IllegalStateException();
         }
+
     }
 
     record NullAccessorRecord(String s, String t) {
+
         public NullAccessorRecord(String s, String t) {
             this.s = s;
             this.t = t + "x";
@@ -231,9 +236,11 @@ class RecordsTest {
         public String s() {
             throw new NullPointerException();
         }
+
     }
 
     record StaticFieldRecord(int i, String s) {
+
         private static final int X = 0;
     }