+++ equalsverifier-core/src/test/java/nl/jqno/equalsverifier/internal/instantiation/vintage/prefabvalues/factories/AbstractGenericFactoryTest.java
@@ -27,7 +27,7 @@ class AbstractGenericFactoryTest {
                     TypeTag tag,
                     VintageValueProvider valueProvider,
                     LinkedHashSet<TypeTag> typeStack) {
-                return Tuple.of("red", "blue", new String("red"));
+                return new Tuple<>("red", "blue", new String("red"));
             }
         };
     }